services/gamification-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
motor==3.3.2
pymongo==4.6.0
redis==5.0.1
celery==5.3.4
python-dotenv==1.0.0


